% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee_kc.R
\name{conumee_kc}
\alias{conumee_kc}
\title{Set of functions to obtain Kc from reference copy number state.}
\usage{
conumee_kc(
  ID,
  ss,
  ref_genes = "paper",
  conumee.folder = "analysis/CONUMEE",
  seg.folder = "Segments",
  log2r.folder = "log2r",
  K = list(Amp10 = 3.666485, Amp = 1.495666, Gains = 1.096756, HetLoss = -1.887569,
    HomDel = -5.124848)
)
}
\arguments{
\item{ID}{ID inside Sample sheet and in the Sample_Name.}

\item{K}{Precalculated constants for each cn state.}

\item{sammp}{identifier for your Kc reference data used in the output names.}
}
\description{
Set of functions to obtain Kc from reference copy number state.
}
\examples{
data("ss")
make_Kc(ss=ss,fname="test",cncols=c("Amp","Amp10","Gains"))
}
