% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{idats_folder}
\alias{idats_folder}
\title{generate champ folder with all idats}
\usage{
idats_folder(targets, folder)
}
\arguments{
\item{targets}{data.frame representing valid targets file as is
used within the minfi package}
}
\value{
folder with idats
}
\description{
prepare idats folder and targets csv in parallel using foreach
}
\author{
izar de Villasante
}
