% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kc_get.R
\name{Kc_get}
\alias{Kc_get}
\title{Set of functions to obtain Kc from reference copy number state.}
\usage{
Kc_get(
  ID,
  ss,
  conumee.folder = "analysis/CONUMEE",
  seg.folder = "Segments",
  log2r.folder = "log2r",
  K = list(Amp10 = 3.666485, Amp = 1.495666, Gains = 1.096756, HetLoss = -1.887569,
    HomDel = -5.124848)
)
}
\arguments{
\item{ID}{ID inside Sample sheet and in the Sample_Name.}

\item{ss}{Sample sheet containing purities, cnstate and Sample_Name ids.}

\item{conumee.folder}{Parent working directory where you want to have your results.}

\item{seg.folder}{Subfolder inside results where segment files are saved.
The segments are generated with CONUMEE CNV.segment(CNV.detail(CNV.bin(fit)))
default = "Segments"}

\item{log2r.folder}{Subfolder inside results where log2r values are saved.
log2r values are the log2 ratio of the intensities(_GRN + _RED channel) between the
query and the reference set (control) as returned by CNV.fit function from CONUMEE.
default = "log2r"}

\item{K}{Precalculated constants for each cn state.}
}
\description{
Set of functions to obtain Kc from reference copy number state.
}
\examples{
data("ss")
cna<-Kc_get(ss=ss,ID="TCGA-19-A6J4-01A-11D-A33U-05")
cna
}
