% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_make.R
\name{make_Kc}
\alias{make_Kc}
\title{Set of functions to obtain Kc from reference copy number state.}
\usage{
make_Kc(
  ss,
  fname,
  cncols = c("Amp10", "Amp", "Gains", "Hetloss", "HomDel"),
  conumee.folder = "analysis/CONUMEE",
  seg.folder = "Segments",
  log2r.folder = "log2r"
)
}
\arguments{
\item{ss}{Sample sheet containing purities, cnstate and Sample_Name ids.}

\item{fname}{identifier for your Kc reference data used in the output names.}

\item{cncols}{column names containing reference set of genes for each cn state}

\item{conumee.folder}{Parent working directory where you want to have your results.}

\item{seg.folder}{Subfolder inside results where segment files are saved.
The segments are generated with CONUMEE CNV.segment(CNV.detail(CNV.bin(fit)))
default = "Segments"}

\item{log2r.folder}{Subfolder inside results where log2r values are saved.
log2r values are the log2 ratio of the intensities(_GRN + _RED channel) between the
query and the reference set (control) as returned by CNV.fit function from CONUMEE.
default = "log2r"}
}
\description{
Set of functions to obtain Kc from reference copy number state.
}
\examples{
data("ss")
make_Kc(ss=ss,fname="test",cncols=c("Amp","Amp10","Gains"))
}
