% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\alias{purify}
\alias{queryfy}
\title{Pre-process methylation array data from (raw) .idat files to segment & log2r intensity files calculated by CONUMEE.}
\usage{
pre_process(
  targets,
  purity = T,
  query = T,
  RGset = T,
  out = "analysis/ChAMP/",
  subf = "IDATS/",
  folder = NULL,
  arraytype = "450K"
)

purify(myLoad)

queryfy(myLoad, out = NULL)
}
\arguments{
\item{targets}{A sample sheet with the required fields for ChAMP to load files.}

\item{purity}{Whether or not you want purity imputed by Rfpurify.
Default=TRUE.}

\item{query}{Set to TRUE when you are only interested in imputing purity.
Default=TRUE}

\item{RGset}{Whether you want normalised "RGChannelSet" or
"RGChannelSetExtended" to be saved or not. path=out}

\item{out}{Results folder to store the normalized and clean RGset.
Default = NULL}

\item{subf}{Subfolder inside results folder where IDAT files will be copied.}

\item{folder}{this directory will be created as the combination of out
and subf. if you save the csv file inside this folder it will be used by
minfi::read.450k.sheet.}

\item{arraytype}{Methylation array type}

\item{myLoad}{a RGset as returned by minfi::read.metharray.exp or
ChAMP.load(method=minfi). ChAMP will use this as basedir to load the idats.
default is results/IDATS/.}
}
\value{
Log2r intensities ready to be analysed with conumee.
}
\description{
Pre-process methylation array data from (raw) .idat files to segment & log2r intensity files calculated by CONUMEE.
}
\examples{

data("TrainingSet_Sample_sheet")
ss<-TrainingSet_Sample_sheet[1:56,]
pre_process(ss)
}
