% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_conumee.R
\name{run_conumee}
\alias{run_conumee}
\alias{read_intensity}
\title{Generate segment & log2r intensity files calculated by CONUMEE.}
\usage{
run_conumee(
  intensities,
  anno_file = NULL,
  ctrl_file = "WB",
  Sample_Name = NULL,
  seg.folder = "Segments",
  log2r.folder = "log2r",
  conumee.folder = "analysis/CONUMEE/",
  probeid = "probeid"
)

read_intensity(infile, Sample_Name = NULL, probeid = "probeid")
}
\arguments{
\item{intensities}{dataset with intensities. either dataframe or path to file.
For big datasets '.fst' file is recomended.}

\item{anno_file}{anno file CNV.anno object saved as .rds. default
CNV_Germline_GSITIC2_BROAD_SNP6.merged.151117.hg19.CNV.txt}

\item{ctrl_file}{CNV data object with intensities from the controls group.
must be compressed as '.rds' file format, default uses 96 WB samples.
Default='WB'}

\item{Sample_Name}{Samples to be analysed. If is NULL all columns in input
file will be used. Accepts numbered index and names. Default=NULL}

\item{seg.folder}{Subfolder inside results where segment files are saved.
The segments are generated with CONUMEE CNV.segment(CNV.detail(CNV.bin(fit)))
default = "Segments"}

\item{log2r.folder}{Subfolder inside results where log2r values are saved.
log2r values are the log2 ratio of the intensities(_GRN + _RED channel) between the
query and the reference set (control) as returned by CNV.fit function from CONUMEE.
default = "log2r"}

\item{conumee.folder}{Parent working directory where you want to have your results.}

\item{probeid}{name of column with probe ids.}

\item{infile}{input path to file or object}
}
\value{
Log2r intensities and segment files.
}
\description{
Generate segment & log2r intensity files calculated by CONUMEE.

read intesity from file or object
}
\examples{

#data("anno")
#data("controls")
#library(conumee)
library(SummarizedExperiment)
intensity<-readRDS("./analysis/ChAMP/intensities.rds")
intensity<-conumee::CNV.load(intensity)@intensity
run_conumee(intensities=intensity)

}
