% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA-package.R
\docType{package}
\name{SCNA-package}
\alias{SCNA}
\alias{SCNA-package}
\title{SCNA: SCNA finds copy number alteration from methylation .idat files}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Izar Llaquet de Villasante \email{idevillasante@carrerasresearch.org} (\href{https://orcid.org/0000-0001-7325-0631 }{ORCID})

}
\keyword{internal}
