% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc_make.R
\name{scna}
\alias{scna}
\title{Function to obtain mean intercept and variance from a
reference gene set and a particular copy number state}
\usage{
scna(ID, log2file, segfile, ss, cn, ref_genes = "all")
}
\arguments{
\item{ID}{name of sample. Must be in Sample_Name column.}

\item{log2file}{file with log2r values, which are the fitted log intensity
value against the reference set returned by CNV.fit function from CONUMEE.}

\item{segfile}{file containing segment data from conumee.
The segments are generated with CONUMEE CNV.segment(CNV.detail(CNV.bin(fit)))}

\item{cn}{Sample sheet column name containing reference gene set for that cn.}

\item{ref_genes}{Granges object with subset of genes to use for cnv analysis.
default = "all"}
}
\value{
dataframe with input data for linear model. Int, mean and Var .
}
\description{
Function to obtain mean intercept and variance from a
reference gene set and a particular copy number state
}
\examples{
ss<-data.table::fread("analysis/ChAMP/Sample_Sheet.txt")
res<-scna(ID="TCGA-05-4405-01A-21D-1856-05",ss=ss,ref_genes="paper",cn="Amp",
log2file="analysis/CONUMEE/log2/TCGA-05-4405-01A-21D-1856-05_log2.txt",
segfile="analysis/CONUMEE/Segments_TCGA-05-4405-01A-21D-1856-05.txt")
res
}
